+++ src/fidesops/ops/service/email/email_dispatch_service.py
@@ -6,6 +6,7 @@ from requests import Response
 from sqlalchemy.orm import Session
 
 from fidesops.ops.common_exceptions import EmailDispatchException
+from fidesops.ops.email_templates import get_email_template
 from fidesops.ops.models.email import EmailConfig
 from fidesops.ops.schemas.email.email import (
     EmailActionType,
@@ -15,6 +16,7 @@ from fidesops.ops.schemas.email.email import (
     EmailServiceType,
     SubjectIdentityVerificationBodyParams,
 )
+from fidesops.ops.util.logger import Pii
 
 logger = logging.getLogger(__name__)
 
@@ -22,29 +24,33 @@ logger = logging.getLogger(__name__)
 def dispatch_email(
     db: Session,
     action_type: EmailActionType,
-    to_email: str,
+    to_email: Optional[str],
     email_body_params: Union[SubjectIdentityVerificationBodyParams],
 ) -> None:
+    if not to_email:
+        raise EmailDispatchException("No email supplied.")
     logger.info("Retrieving email config")
     email_config: Optional[EmailConfig] = db.query(EmailConfig).first()
     if not email_config:
         raise EmailDispatchException("No email config found.")
     if not email_config.secrets:
         logger.warning(
-            f"Email secrets not found for config with key: {email_config.key}"
+            "Email secrets not found for config with key: %s", email_config.key
         )
         raise EmailDispatchException(
             f"Email secrets not found for config with key: {email_config.key}"
         )
-    logger.info(f"Building appropriate email template for action type: {action_type}")
+    logger.info("Building appropriate email template for action type: %s", action_type)
     email: EmailForActionType = _build_email(
         action_type=action_type, body_params=email_body_params
     )
     email_service: EmailServiceType = email_config.service_type  # type: ignore
-    logger.info(f"Retrieving appropriate dispatcher for email service: {email_service}")
+    logger.info(
+        "Retrieving appropriate dispatcher for email service: %s", email_service
+    )
     dispatcher: Any = _get_dispatcher_from_config_type(email_service_type=email_service)
     logger.info(
-        f"Starting email dispatch for email service with action type: {action_type}"
+        "Starting email dispatch for email service with action type: %s", action_type
     )
     dispatcher(email_config=email_config, email=email, to_email=to_email)
 
@@ -54,12 +60,17 @@ def _build_email(
     body_params: Union[SubjectIdentityVerificationBodyParams],
 ) -> EmailForActionType:
     if action_type == EmailActionType.SUBJECT_IDENTITY_VERIFICATION:
+        template = get_email_template(action_type)
         return EmailForActionType(
             subject="Your one-time code",
-            # for 1st iteration, below will be replaced with actual template files
-            body=f"<html>Your one-time code is {body_params.access_code}. Hurry! It expires in 10 minutes.</html>",
+            body=template.render(
+                {
+                    "code": body_params.verification_code,
+                    "minutes": body_params.get_verification_code_ttl_minutes(),
+                }
+            ),
         )
-    logger.error(f"Email action type {action_type} is not implemented")
+    logger.error("Email action type %s is not implemented", action_type)
     raise EmailDispatchException(f"Email action type {action_type} is not implemented")
 
 
@@ -97,11 +108,11 @@ def _mailgun_dispatcher(
         )
         if not response.ok:
             logger.error(
-                f"Email failed to send with status code: {response.status_code}"
+                "Email failed to send with status code: %s", response.status_code
             )
             raise EmailDispatchException(
                 f"Email failed to send with status code {response.status_code}"
             )
     except Exception as e:
-        logger.error("Email failed to send: %s", str(e))
+        logger.error("Email failed to send: %s", Pii(str(e)))
         raise EmailDispatchException(f"Email failed to send due to: {e}")